Funcional 29/09/2016

+++ MProjectWeb/src/MProjectWeb/Models/Postgres/repositorios_usuarios.cs
@@ -5,7 +5,7 @@ namespace MProjectWeb.Models.postgres
 {
     public partial class repositorios_usuarios
     {
-        public int id_usuario { get; set; }
+        public long id_usuario { get; set; }
         public string ruta_repositorio { get; set; }
 
         public virtual usuarios id_usuarioNavigation { get; set; }
